color: refactor the attribute handling
Refactor the common code out of two attribute handling functions.
Create a Mapping table, AttributeNames.

- parse_attr_spec()
- parse_color_pair()
